+++ rocAL/rocAL/rocAL_hip/box_encoder_hip.cpp
@@ -23,7 +23,7 @@ THE SOFTWARE.
 #include "box_encoder_hip.h"
 
 // for both device and host
-__host__ __device__ inline float4 ToBoxCenterWH(const float4 &box) {
+__host__ __device__ inline double4 ToBoxCenterWH(const double4 &box) {
     return {
       0.5f * (box.x + box.z),
       0.5f * (box.y + box.w),
@@ -31,9 +31,9 @@ __host__ __device__ inline float4 ToBoxCenterWH(const float4 &box) {
       box.w - box.y};
 }
 
-__device__ float4 MatchOffsets(float4 box, float4 anchor, const float *means, const float *inv_stds, float scale) {
+__device__ double4 MatchOffsets(double4 box, double4 anchor, const double *means, const double *inv_stds, double scale) {
 
-    float4 box_out;
+    double4 box_out;
     box.x *= scale; box.y *= scale; box.z *= scale; box.w *= scale;
     anchor.x *= scale; anchor.y *= scale; anchor.z *= scale; anchor.w *= scale;
 
@@ -46,21 +46,21 @@ __device__ float4 MatchOffsets(float4 box, float4 anchor, const float *means, co
 }
 
 
-__device__ __forceinline__ float CalculateIou(const float4 &b1, const float4 &b2) {
-    float l = fmaxf(b1.x, b2.x);
-    float t = fmaxf(b1.y, b2.y);
-    float r = fminf(b1.z, b2.z);
-    float b = fminf(b1.w, b2.w);
-    float first = fmaxf(r - l, 0.0f);
-    float second = fmaxf(b - t, 0.0f);
-    volatile float intersection = first * second;
-    volatile float area1 = (b1.w - b1.y) * (b1.z - b1.x);
-    volatile float area2 = (b2.w - b2.y) * (b2.z - b2.x);
+__device__ __forceinline__ double CalculateIou(const double4 &b1, const double4 &b2) {
+    double l = fmaxf(b1.x, b2.x);
+    double t = fmaxf(b1.y, b2.y);
+    double r = fminf(b1.z, b2.z);
+    double b = fminf(b1.w, b2.w);
+    double first = fmaxf(r - l, 0.0f);
+    double second = fmaxf(b - t, 0.0f);
+    volatile double intersection = first * second;
+    volatile double area1 = (b1.w - b1.y) * (b1.z - b1.x);
+    volatile double area2 = (b2.w - b2.y) * (b2.z - b2.x);
 
     return intersection / (area1 + area2 - intersection);
 }
 
-__device__ inline void FindBestMatch(const int N, volatile float *vals, volatile int *idx) {
+__device__ inline void FindBestMatch(const int N, volatile double *vals, volatile int *idx) {
   for (unsigned int stride = blockDim.x / 2; stride > 0; stride >>= 1) {
     if (threadIdx.x < stride) {
       if (vals[threadIdx.x] <= vals[threadIdx.x + stride]) {
@@ -76,15 +76,15 @@ __device__ inline void FindBestMatch(const int N, volatile float *vals, volatile
   }
 }
 
-__device__ void WriteMatchesToOutput(unsigned int anchor_count, float criteria, int *labels_out, const int *labels_in,
-                                    float4 *boxes_out, const float4 *boxes_in, volatile int *best_box_idx, volatile float *best_box_iou,
-                                    bool offset, const float* means, const float* inv_stds, float scale, const float4 *anchors_as_cwh) {
+__device__ void WriteMatchesToOutput(unsigned int anchor_count, double criteria, int *labels_out, const int *labels_in,
+                                    double4 *boxes_out, const double4 *boxes_in, volatile int *best_box_idx, volatile double *best_box_iou,
+                                    bool offset, const double* means, const double* inv_stds, double scale, const double4 *anchors_as_cwh) {
 
     for (unsigned int anchor = threadIdx.x; anchor < anchor_count; anchor += blockDim.x) {
         if (best_box_iou[anchor] > criteria) {
             int box_idx = best_box_idx[anchor];
             labels_out[anchor] = labels_in[box_idx];
-            float4 box = boxes_in[box_idx];
+            double4 box = boxes_in[box_idx];
 
             if (!offset)
               boxes_out[anchor] = ToBoxCenterWH(box);
@@ -95,14 +95,14 @@ __device__ void WriteMatchesToOutput(unsigned int anchor_count, float criteria, 
 }
 
 
-__device__ void MatchBoxWithAnchors(const float4 &box, const int box_idx, unsigned int anchor_count, const float4 *anchors,
-                                    volatile int *best_anchor_idx_buf, volatile float *best_anchor_iou_buf,
-                                    volatile int *best_box_idx, volatile float *best_box_iou) {
-    float best_anchor_iou = -1.0f;
+__device__ void MatchBoxWithAnchors(const double4 &box, const int box_idx, unsigned int anchor_count, const double4 *anchors,
+                                    volatile int *best_anchor_idx_buf, volatile double *best_anchor_iou_buf,
+                                    volatile int *best_box_idx, volatile double *best_box_iou) {
+    double best_anchor_iou = -1.0f;
     int best_anchor_idx = -1;
 
     for (unsigned int anchor = threadIdx.x; anchor < anchor_count; anchor += blockDim.x) {
-      float new_val = CalculateIou(box, anchors[anchor]);
+      double new_val = CalculateIou(box, anchors[anchor]);
 
       if (new_val >= best_anchor_iou) {
           best_anchor_iou = new_val;
@@ -121,9 +121,9 @@ __device__ void MatchBoxWithAnchors(const float4 &box, const int box_idx, unsign
 
 template <int BLOCK_SIZE>
 __global__ void __attribute__((visibility("default")))
-BoxEncode(const BoxEncoderSampleDesc *samples, const int anchor_cnt, const float4 *anchors,
-          const float criteria, int *box_idx_buffer, float *box_iou_buffer, bool offset,
-          const float *means, const float *inv_stds, float scale, const float4 *anchors_as_cwh) {
+BoxEncode(const BoxEncoderSampleDesc *samples, const int anchor_cnt, const double4 *anchors,
+          const double criteria, int *box_idx_buffer, double *box_iou_buffer, bool offset,
+          const double *means, const double *inv_stds, double scale, const double4 *anchors_as_cwh) {
     // printf("\n In Box Encoder kernel");
     // printf("\n invs_stds1:%f",inv_stds[0]);
 
@@ -131,10 +131,10 @@ BoxEncode(const BoxEncoderSampleDesc *samples, const int anchor_cnt, const float
     const auto &sample = samples[sample_idx];
 
     __shared__ volatile int best_anchor_idx_buf[BLOCK_SIZE];
-    __shared__ volatile float best_anchor_iou_buf[BLOCK_SIZE];
+    __shared__ volatile double best_anchor_iou_buf[BLOCK_SIZE];
 
     volatile int *best_box_idx = box_idx_buffer + sample_idx * anchor_cnt;
-    volatile float *best_box_iou = box_iou_buffer + sample_idx * anchor_cnt;
+    volatile double *best_box_iou = box_iou_buffer + sample_idx * anchor_cnt;
 
     for (int box_idx = 0; box_idx < sample.in_box_count; ++box_idx) {
       MatchBoxWithAnchors(
@@ -179,16 +179,16 @@ BoxEncode(const BoxEncoderSampleDesc *samples, const int anchor_cnt, const float
 
 }
 
-void BoxEncoderGpu::prepare_anchors(const std::vector<float> &anchors) {
+void BoxEncoderGpu::prepare_anchors(const std::vector<double> &anchors) {
 
     if ((anchors.size() % 4) != 0)
         THROW("BoxEncoderGpu anchors not a multiple of 4");
 
     int anchor_count = anchors.size() / 4;
-    int anchor_data_size = anchor_count * 4 * sizeof(float);
-    auto anchors_data_cpu = reinterpret_cast<const float4 *>(anchors.data());
+    int anchor_data_size = anchor_count * 4 * sizeof(double);
+    auto anchors_data_cpu = reinterpret_cast<const double4 *>(anchors.data());
 
-    std::vector<float4> anchors_as_center_wh(anchor_count);
+    std::vector<double4> anchors_as_center_wh(anchor_count);
     for (unsigned int anchor = 0; anchor < anchor_count; ++anchor)
       anchors_as_center_wh[anchor] = ToBoxCenterWH(anchors_data_cpu[anchor]);
 
@@ -196,28 +196,28 @@ void BoxEncoderGpu::prepare_anchors(const std::vector<float> &anchors) {
     HIP_ERROR_CHECK_STATUS(hipMemcpy((void *)_anchors_as_center_wh_data_dev, anchors_as_center_wh.data(), anchor_data_size, hipMemcpyHostToDevice));
 }
 
-void BoxEncoderGpu::prepare_mean_std(const std::vector<float> &means, const std::vector<float> &stds) {
+void BoxEncoderGpu::prepare_mean_std(const std::vector<double> &means, const std::vector<double> &stds) {
 
-    int data_size = 4 * sizeof(float);
-    auto means_data_cpu = reinterpret_cast<const float *>(means.data());
-    auto stds_data_cpu = reinterpret_cast<const float *>(stds.data());
+    int data_size = 4 * sizeof(double);
+    auto means_data_cpu = reinterpret_cast<const double *>(means.data());
+    auto stds_data_cpu = reinterpret_cast<const double *>(stds.data());
 
     HIP_ERROR_CHECK_STATUS(hipMemcpy((void *)_means_dev, means_data_cpu, data_size, hipMemcpyHostToDevice));
     HIP_ERROR_CHECK_STATUS(hipMemcpy((void *)_stds_dev, stds_data_cpu, data_size, hipMemcpyHostToDevice));
 }
 
-void BoxEncoderGpu::WriteAnchorsToOutput(float* encoded_boxes) {
+void BoxEncoderGpu::WriteAnchorsToOutput(double* encoded_boxes) {
   // Device -> device copy for all the samples
   for (int i=0; i<_cur_batch_size; i++) {
     HIP_ERROR_CHECK_STATUS(hipMemcpyDtoDAsync((void *)(encoded_boxes + i*_anchor_count*4), _anchors_as_center_wh_data_dev,
-                                            _anchor_count * 4 * sizeof(float), _stream));
+                                            _anchor_count * 4 * sizeof(double), _stream));
   }
 }
 
 
-std::pair<int *, float *> BoxEncoderGpu::ResetBuffers() {
+std::pair<int *, double *> BoxEncoderGpu::ResetBuffers() {
     HIP_ERROR_CHECK_STATUS(hipMemsetAsync(_best_box_idx_dev, 0, _cur_batch_size * _anchor_count * sizeof(int), _stream));
-    HIP_ERROR_CHECK_STATUS(hipMemsetAsync(_best_box_iou_dev, 0, _cur_batch_size * _anchor_count * sizeof(float), _stream));
+    HIP_ERROR_CHECK_STATUS(hipMemsetAsync(_best_box_iou_dev, 0, _cur_batch_size * _anchor_count * sizeof(double), _stream));
     return std::make_pair(_best_box_idx_dev, _best_box_iou_dev);
 }
 
@@ -225,12 +225,12 @@ void BoxEncoderGpu::ResetLabels(int *encoded_labels_out) {
     HIP_ERROR_CHECK_STATUS(hipMemsetAsync((void *)encoded_labels_out, 0, _cur_batch_size *_anchor_count * sizeof(int), _stream));
 }
 
-void BoxEncoderGpu::ClearOutput(float* encoded_boxes) {
-    HIP_ERROR_CHECK_STATUS(hipMemsetAsync(encoded_boxes, 0, _cur_batch_size*_anchor_count * 4 * sizeof(float), _stream));
+void BoxEncoderGpu::ClearOutput(double* encoded_boxes) {
+    HIP_ERROR_CHECK_STATUS(hipMemsetAsync(encoded_boxes, 0, _cur_batch_size*_anchor_count * 4 * sizeof(double), _stream));
 }
 
 
-void BoxEncoderGpu::Run(pMetaDataBatch full_batch_meta_data, float *encoded_boxes_data, int *encoded_labels_data) {
+void BoxEncoderGpu::Run(pMetaDataBatch full_batch_meta_data, double *encoded_boxes_data, int *encoded_labels_data) {
 
     if (_cur_batch_size != full_batch_meta_data->size() || (_cur_batch_size <=0))
         THROW("BoxEncoderGpu::Run Invalid input metadata");
@@ -244,22 +244,22 @@ void BoxEncoderGpu::Run(pMetaDataBatch full_batch_meta_data, float *encoded_boxe
     }
     if (total_num_boxes > MAX_NUM_BOXES_TOTAL)
         THROW("BoxEncoderGpu::Run total_num_boxes exceeds max");
-    float *boxes_in_temp = _boxes_in_dev; int *labels_in_temp = _labels_in_dev;
+    double *boxes_in_temp = _boxes_in_dev; int *labels_in_temp = _labels_in_dev;
     for (int sample_idx = 0; sample_idx < _cur_batch_size; sample_idx++) {
         auto sample = &_samples_host_buf[sample_idx];
         //sample->in_box_count = full_batch_meta_data->get_bb_labels_batch()[sample_idx].size();
-        HIP_ERROR_CHECK_STATUS( hipMemcpyHtoDAsync((void *)boxes_in_temp, full_batch_meta_data->get_bb_cords_batch()[sample_idx].data(), sample->in_box_count*sizeof(float)*4, _stream));
+        HIP_ERROR_CHECK_STATUS( hipMemcpyHtoDAsync((void *)boxes_in_temp, full_batch_meta_data->get_bb_cords_batch()[sample_idx].data(), sample->in_box_count*sizeof(double)*4, _stream));
         HIP_ERROR_CHECK_STATUS( hipMemcpyHtoDAsync((void *)labels_in_temp, full_batch_meta_data->get_bb_labels_batch()[sample_idx].data(), sample->in_box_count*sizeof(int), _stream));
-        sample->boxes_in = reinterpret_cast<const float4 *>(boxes_in_temp);
+        sample->boxes_in = reinterpret_cast<const double4 *>(boxes_in_temp);
         sample->labels_in = reinterpret_cast<const int *>(labels_in_temp);
-        sample->boxes_out = reinterpret_cast<float4 *>(encoded_boxes_data + sample_idx*_anchor_count*4);
+        sample->boxes_out = reinterpret_cast<double4 *>(encoded_boxes_data + sample_idx*_anchor_count*4);
         sample->labels_out = reinterpret_cast<int *>(encoded_labels_data + sample_idx*_anchor_count);
         boxes_in_temp += sample->in_box_count*4;
         labels_in_temp += sample->in_box_count;
         _output_shape.push_back(std::vector<size_t>(1,_anchor_count));
     }
-    const auto means_data = reinterpret_cast<const float *>(_means.data());
-    const auto stds_data = reinterpret_cast<const float *>(_stds.data());
+    const auto means_data = reinterpret_cast<const double *>(_means.data());
+    const auto stds_data = reinterpret_cast<const double *>(_stds.data());
 
     // std::cerr<<"stds_data1:"<<stds_data[0]<<"stds_data1:"<<stds_data[1]<<"stds_data2:"<<stds_data[2]<<"stds_data3:"<<stds_data[3];
