+++ rocAL/rocAL_pybind/rocal_pybind.cpp
@@ -172,16 +172,16 @@ namespace rocal{
         return py::cast<py::none>(Py_None);
     }
 
-    std::pair<py::array_t<float>, py::array_t<int>>  wrapper_get_encoded_bbox_label(RocalContext context, int batch_size, int num_anchors)
+    std::pair<py::array_t<double>, py::array_t<int>>  wrapper_get_encoded_bbox_label(RocalContext context, int batch_size, int num_anchors)
     {
-        float* bboxes_buf_ptr; int* labels_buf_ptr;
+        double* bboxes_buf_ptr; int* labels_buf_ptr;
         // call pure C++ function
         rocalGetEncodedBoxesAndLables(context, &bboxes_buf_ptr, &labels_buf_ptr, num_anchors*batch_size);
         // create numpy arrays for boxes and labels tensor from the returned ptr
         // no need to free the memory as this is freed by c++ lib
-        py::array_t<float> bboxes_array = py::array_t<float>(
+        py::array_t<double> bboxes_array = py::array_t<double>(
                                                           {batch_size, num_anchors, 4},
-                                                          {4*sizeof(float)*num_anchors, 4*sizeof(float), sizeof(float)},
+                                                          {4*sizeof(double)*num_anchors, 4*sizeof(double), sizeof(double)},
                                                           bboxes_buf_ptr,
                                                           py::cast<py::none>(Py_None));
         py::array_t<int> labels_array = py::array_t<int>(
